+++ packages/doars-fetch/src/DoarsFetch.js
@@ -7,12 +7,20 @@ import createFetchContext from "./contexts/fetch.js";
 import createFetchDirective from "./directives/fetch.js";
 
 /**
+ * @typedef {Object} ResponseParser
+ * @property {string[]} types - The types of responses the parser can handle.
+ * @property {(response: Response, type: string) => Promise<any>} parser - The function to parse the response.
+ */
+
+/**
  * @typedef DoarsFetchOptions
  * @type {object}
  * @property {string} [fetchContextName] - The name of the fetch context.
  * @property {boolean} [fetchDirectiveEvaluate] - If set to false the fetch directive's value is read as a string literal instead of an expression to process.
  * @property {string} [fetchDirectiveName] - The name of the fetch directive.
  * @property {object} [fetchOptions] - Default fetch options to use, the options object provided when calling fetch will be merged with this default.
+ * @property {boolean} [fetchAutoParse=true] - Whether to automatically parse the response based on content type. Enabled by default.
+ * @property {ResponseParser[]} [fetchParsers] - Custom parsers to use in addition to built-in ones. Useful for adding support for YAML, TOML, CSV, etc.
  * @property {string|false} [intersectionEvent] - The name of the intersect special event listener. To disable the event from ever triggering set this option to false.
  * @property {HTMLElement} [intersectionRoot] - The element to be used as the viewport for checking the visibility of the elements.
  * @property {string} [intersectionMargin] - Margin around the root.
@@ -33,6 +41,8 @@ export default function (library, options = null) {
 			fetchDirectiveEvaluate: true,
 			fetchDirectiveName: "fetch",
 			fetchOptions: {},
+			fetchAutoParse: true,
+			fetchParsers: [],
 
 			intersectionEvent: "intersect",
 			intersectionRoot: null,
@@ -43,6 +53,7 @@ export default function (library, options = null) {
 		},
 		options,
 	);
+	// Backwards compatibility assign.
 	if (options.defaultInit) {
 		Object.assign(options.fetchOptions, options.defaultInit);
 	}